requirements.txt
duckdb
postgres
tabulate
kagglehub
matplotlib
python-dotenv